move controllerconfig out of componentconfig
